Bug Fix: Making the programme more OOP-Oriented
