+++ data/pixel.txt
@@ -1,7 +1,1 @@
-<<<<<<< HEAD
-D | Not Done | CS2103 assignment | tomorrow
-=======
-D | Not Done | asdsad| 
-T | Not Done | dfd s| 
-E | Not Done | fireworks| 
->>>>>>> f8433334b6723da785457d67f764e8cb73bbac74
+D | Not Done | dfgfdggd | d
